Moving some files from lib to the battle subdirectory.
