+++ AI/StupidAI/StupidAI.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../lib/BattleHex.h"
+#include "../../lib/battle/BattleHex.h"
 
 class CStupidAI : public CBattleGameInterface
 {
