+++ CCallback.cpp
@@ -18,7 +18,7 @@
 #include "lib/GameConstants.h"
 #include "lib/CPlayerState.h"
 #include "lib/UnlockGuard.h"
-#include "lib/BattleInfo.h"
+#include "lib/battle/BattleInfo.h"
 
 /*
  * CCallback.cpp, part of VCMI engine
