+++ client/battle/CBattleAnimations.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../lib/BattleHex.h"
+#include "../../lib/battle/BattleHex.h"
 #include "../widgets/Images.h"
 
 class CBattleInterface;
