+++ lib/battle/BattleHex.cpp
