+++ lib/battle/BattleInfo.cpp
