+++ lib/battle/BattleInfo.h
