+++ lib/battle/CObstacleInstance.cpp
