+++ lib/battle/CObstacleInstance.h
