+++ lib/registerTypes/RegisterTypes.h
@@ -11,7 +11,7 @@
 #include "../mapObjects/CRewardableConstructor.h"
 #include "../mapObjects/CommonConstructors.h"
 #include "../mapObjects/MapObjects.h"
-#include "../CObstacleInstance.h"
+#include "../battle/CObstacleInstance.h"
 
 /*
  * RegisterTypes.h, part of VCMI engine
