+++ lib/spells/CSpellHandler.cpp
@@ -23,7 +23,7 @@
 #include "../StringConstants.h"
 
 #include "../CStack.h"
-#include "../BattleInfo.h"
+#include "../battle/BattleInfo.h"
 #include "../battle/CBattleInfoCallback.h"
 #include "../CGameState.h" //todo: remove
