+++ server/NetPacksServer.cpp
@@ -6,8 +6,8 @@
 #include "../lib/mapping/CMap.h"
 #include "../lib/CGameState.h"
 #include "../lib/CStack.h"
-#include "../lib/BattleInfo.h"
-#include "../lib/BattleAction.h"
+#include "../lib/battle/BattleInfo.h"
+#include "../lib/battle/BattleAction.h"
 #include "../lib/serializer/Connection.h"
