+++ test/Battlefield.cpp
@@ -9,7 +9,7 @@
  */
 #include "StdInc.h"
 #include <boost/test/unit_test.hpp>
-#include "../lib/BattleHex.h"
+#include "../lib/battle/BattleHex.h"
 BOOST_AUTO_TEST_SUITE(BattlefieldHex_Suite)
 
 BOOST_AUTO_TEST_CASE(getNeighbouringTiles)
